requirements.txt
aiofiles==0.4.0
git+https://github.com/arugifa/cms@dev#egg=arugifa-cms
git+https://github.com/arugifa/py-toolbox@dev#egg=arugifa-toolbox
factory-boy==2.11.1
flask==1.0.2
flask-sqlalchemy==2.4.0
frozen-flask==0.15
lxml[cssselect]==4.3.3
openstacksdk==0.38.0
sortedcontainers==2.1.0
sqlalchemy==1.3.3
